Load all plugins based on the accounts on startup.
You can avoid this by setting Enabled=false in kopeterc for the given
account.

I'll probably add a command line switch or some GUI for it later for developers
but at least we all agreed that for end users this is the way to go.

Note that this overlaps with the normal plugin loader (it may attempt to
load a plugin twice), but the plugin loader is smart enough to handle that,
so it's not a problem for now. I'll remove the loading of protocols through
::loadAllPlugins through a kconf_update script that simply removes the keys
when the GUI is gone too.

Whatever we do in the GUI or the command line, this is the backend part that
was NOT controversial, so now we only need to discuss the remaining part :)

CCMAIL: kopete-devel@kde.org

svn path=/trunk/kdenetwork/kopete/; revision=260607