Do a regex replace on all instances of contactDisplayNames in new messages to facilitae the right click menus anywhere a displayname appears.
I never did this originally because I thought it would be a horrible performer,
however after experimenting with it in chats with even over 650 people in them,
I have no lag at all! QRegExp is very efficient.

However I would appreciate someone with a slower PC testing this as well.

If it is deemed too slow I can always add a check in the function and only
do this replacement if there are less than a certain number of chat members
present.

svn path=/trunk/kdenetwork/kopete/; revision=261176