Correct the colouring of channel and server icons.
The reason for this was that the base icon and the icon used for the status overlay were one and the same.  After colourising the base, the overlay is drawn onto it.  If they are identical, any colourisation will be completely covered.

Note that it is never a good idea to take your overlay icons (specified in KopeteOnlineStatus ctors) from the same set of images that comprise your base icons.  If you don't want an overlay, use QString::null for the 5th argument to the KOS ctor.

CCMAIL: 66224-done@bugs.kde.org

svn path=/trunk/kdenetwork/kopete/; revision=262447