Port to QPtrList<KAction>* as discussed on list.
NOTE: Because QPtrList is not a QObject, the auto-garbage collection
hacking we used to have by re-parenting the KActions is no longer. I have
accounted for this everywhere in the code EXCEPT for one Jabber
KAction, that is still leaking because there was some fair amounts
of code in there I didn't quite understand, so I left it for the
Jabber guys to take care of

Note also, if anyone uses this function for something in the future,
you need to manually delete the QPtrList when you're done with it.

CCMAIl:kopete-devel@kde.org

svn path=/trunk/kdenetwork/kopete/; revision=263343