Fix bug 64918. Here's the whole scoop:
When clicking on a metacontact, the isReachable function is called, and the
same error message that is given in oscar should now be given in all
protocols. There is a catch however. If you have multiple protocol contacts
in a meta contact (i.e. for one meta contact you have a yahoo contact, an msn
contact, etc.) and you click on the little protocol icon to message them using
a certain protocol, isReachable() doesn't appear to be called and you still get
the same behavior.  This is a bug in libkopete and I will file a new bug report
for that shortly.

I consider this bug fixed. _Please_ do not reopen this bug unless I missed one.
(i.e. you have one contact per metacontact and don't get the appropriate error
message box). And if I missed one, I expect to be screamed and shouted at. :)

svn path=/trunk/kdenetwork/kopete/; revision=263853