This commit fixes the conneciton in MSN.
The MSN Server just changed a bit. It now does not accept anymore an old command I used.
This is now fixed.
Anyway, Should we backport this to the 0_7_branch  and release Kopete 0.7.4 (or 0.7.3a) ?

CCMAIL: kopete-devel@kde.org
CCMAIL: 68204-done@bugs.kde.org

Note: It's still possible to connect with the old version if you force kopete to download the contactlist from server:
i.e: moving a MSN contact in another group (when you are offline)
Or editing the kopeterc file and reset the sirial to 0

svn path=/trunk/kdenetwork/kopete/; revision=267072