Fix bug 61223. Instead of just disregarding the fact that the client needs authorization when we add them, resend the add request again, telling the ICQ server that we need authorization from the person on the other end. The authorization message that is sent is blank to avoid conflict with the string freeze.
Please note that contacts are still shown offline until they authorize your
request to add them to your contact list. This is inline with the behavior
of the official ICQ client.

Also, contacts that are currently on your list are not affected by this patch.
You can right-click on the ICQ flower for that contact and click the
"Request Authorization" button if it is enabled. If that button is greyed out,
you will have to remove the contact from your contact list and readd them again
after compiling Kopete with this patch.

CCMAIL: 61223-done@bugs.kde.org

svn path=/trunk/kdenetwork/kopete/; revision=273779