Test application for the emoticon engine. This is one of the very few parts of Kopete that can be easily tested using automated testing, and this app should help there.
Whenever you change something in the emoticon engine, run the test and make
sure at least the 'Known Good' emoticons aren't broken.

To build, run 'make check' in the tests dir. The tests are hardcoded for
the 'default' emoticon theme, but that shouldn't spoil the fun, at least not
until somebody changes that theme :)

If you have more tests, feel free to add them to the appropriate entry, it's
good to have a lot of tests so we know what to fix in later Kopete versions.

I don't have a Kopete 0.7.x around anymore, so I didn't fill in the baseline.
If someone can run the test against 0.7 and move the tests that succeed there
to the baseline that would be awesome. Especially to see what fails now but
succeeded back then.

CCMAIL: kopete-devel@kde.org

svn path=/trunk/kdenetwork/kopete/; revision=275625