Make the MSN buddy icon not stretch the entire toolbar. This is a slightly cleaned up version of the patch I posted to kopete-devel and which Olivier approved.
CCMAIL: 66398-fixed@bugs.kde.org

As for the FIXMEs, if anyone knows how to determine the height of a toolbar
and adjust the icon to it, that would be nice. Switching the KWidgetAction
with another type of KAction is also fine, but I can't think of a working
combination off-hand.

Olivier, to fix the code duplication FIXME, would it be possible to call
slotDisplayPictureChanged from the constructor too? The two code paths are
not 100% identical since the slot checks for AutoDownloadPicture and the
c'tor does not. Is that safe from the constructor?

svn path=/trunk/kdenetwork/kopete/; revision=275888