Can't beleive no one'es reported / noticed this one yet, it was for sure a showstopper.
Deleting contacts was not happening in the right sequence, so if you
removed a contact from your list, he was also removed from all the chats
he was involved in, even if KopeteContact::slotDeleteContact did not dleete
him.

This is because slotDeleteContact was being called, but then
the KopeteMetaContact was being deleted, which deleted him anyway! (because
he was still a child QObject )

This re-orderes things, so now when you delete a MC, first all the child
contacts are removed from it, *then* slotDeleteContact is called on them.

svn path=/trunk/kdenetwork/kopete/; revision=278464