Text Codec selection in IRC.
I know this breaks string freeze, but I was told by several i18n members
that this is a showstopper for non English locales, so has to be fixed.

I could not use the ICQ string because of the : after Encoding (It's in
a widget, so it has "Encoding:" beside a dropdown). I did, however, use
a string from the KSirc pot. I don't know if ths can be shared or not,
I know nothing about i18n, but it was the best bet I had.

The new string is "&Encoding"

CCMAIL: kde-i18n-doc@kde.org
CCMAIL: kopete-devel@kde.org

svn path=/trunk/kdenetwork/kopete/; revision=279264