Add going Offline to the list of changes that should be suppressed; use the suppressions also when displaying of events in the chat window is enabled. Don't display an event for Connecting, there's Offline and the various Online/Away states for that.
CCMAIL: 71076-done@bugs.kde.org

The fix is dependent on the suppression of status notifications, so it
can't easily get backported. It also adds a new string, but that is in
theory avoidable. However, the bug is fairly minor, so I personally
would like to keep this in HEAD only for Kopete 0.9.

svn path=/trunk/kdenetwork/kopete/; revision=285900