Make a contact instance honour a default resource being supplied to it upon creation.
For those listed in the bug report: Peter, I don't know if it makes sense to
only subscribe to a certain resource of a contact? Is that actually allowed?
The Jabber protocol does not forbid it, nor does it allow it. Right now, the
code will allow you to "subscribe" to a resource by adding this specific
resource to the contact list, but it will always stay offline.

At least Psi also adds a normal contact of the form <user@domain> to the
list, which all reply messages will then originate from. Any input is
greatly appreciated.
CCMAIL: 63425@bugs.kde.org

svn path=/trunk/kdenetwork/kopete/; revision=294938