Revert a part of the commit from Jason  at 2004/1/10 ( kopetecontactlist.cpp rev 1.117 and kopetemetacontact.cpp rev 1.197)
It was a try to fix a bug in IRC, but the result is to broke every protocol, even IRC.
That's causing thet KopeteContact::metaContact() can retrun a dangling pointer and cause crash while deleting msn metacontact while offline, or irc contact chen they are in a chat.

IRC still need to be fixed.  a possibility: in IRCContact::slotDeleteContact recreate a new temporary metacontact and move the contact in.

CCMAIL:  jason@keirstead.org

svn path=/trunk/kdenetwork/kopete/; revision=295813