I've made sure the keepalive code works in HEAD (which it does), and these pieces of code are to help with the duplicate login bug. There is a two minute delay before the yahoo protocol determines that it's been disconnected. If after being disconnected for approximately two minutes, the yahoo protocol still shows offline, please open a different bug.
For the duplicate login bug, I check to see if we've been disconnected by the
keepalive mechanism and suppress duplicate login warning (because it's not
really a duplicate login, the yahoo servers just don't think we've disconnected
yet)

Please note that this is the best way that I can think of within the limits of
the protocol. Yahoo is not nearly as accurate about detecting disconnects as
the MSN plugin is, and it's due to the way the YMSG protocol is designed.

(What a long commit message for so little code) :)

CCMAIL: 67287-done@bugs.kde.org
CCMAIL: 70311-done@bugs.kde.org

svn path=/trunk/kdenetwork/kopete/; revision=297472