Allow accounts to connect as something other than online, part two:
Reworked the connection stuff to get rid of the enum and use KopeteOnlineStatus
instead. This allows anyone to connect as anything as long as it's represented
as a KopeteOnlineStatus object somehow, so the sky's pretty much the limit here.

Review is welcome. I made sure it compiled (and linked), but that's about it.
Nothing uses it yet.

CCMAIL: kopete-devel@kde.org

svn path=/trunk/kdenetwork/kopete/; revision=300578