Kopete now supports KIMProxy by implementing KIMIface, the generic IM interface.
This allows any app using KIMProxy to dynamically get IM information from Kopete.

The implementation takes the form of another DCOP interface, which reduces the impact on Kopete.  All the changes are in KIMIfaceImpl, which works with Kopete through its public API.

The desktop file is changed to specify that Kopete implements DCOP/InstantMessenger, so that KIMProxy and KControl's Component Chooser can locate Kopete as a messaging service provider.

Inclusion of kimiface.h is a temporary compatibility measure until kimproxy becomes part of kdelibs/interfaces, so I will have to take care that this remains synced with kdenonbeta/kimproxy.

svn path=/trunk/kdenetwork/kopete/; revision=304235