Add Custom Notifications aka Buddy Pounces to Kopete.  This lets you override the Kopete-wide notifications with ones that are specific to individual contacts and that may only fire once.
To use, see the Properties dialog accessible from each contact's context menu.

Implementation:  The events are stored in a similar way to KopetePluginDataObject properties, but relate to metacontacts not plugins.  When an event is triggered, I have hacked KopeteNotifyClient further to check for custom notifications first and process them, overriding common notifications by default.  The notifications are stored in the contactlist.xml file.

Todo:

* add kconf_update script to rename users' customised kopete.eventsrc (I had to rename our events so that the custom notifications dialog can filter those out not relevant to contacts).
* let groups have custom notifications too (mostly ui plumbing needed here).
* let us get the audio file for a notification from KABC's field for this.

svn path=/trunk/kdenetwork/kopete/; revision=326127