Temporary workaround.
Let the user talk to contact not on his contactlist. contact which blocked him, or in
the case to chat session is sitll open.  Let the user talk to contact which are not in the contactlist.

Yes, this is not the correct way to do. Maybe i'll try to propose a better fix later.

svn path=/trunk/kdenetwork/kopete/; revision=327923