Divide per two the time needed to log a message.
I designed the history code to load the history file only once.
But unfortunately, i saved it to a copy of the map. So it was not realy keeped in memory, and the logger still loaded the document each time.

CCMAIL: 65982@bugs.kde.org
Actualy, the file is still saved on the disk for each message. I'll add a timer to save less often.

svn path=/trunk/kdenetwork/kopete/; revision=329865