Make it possible to do incremental search again.
Justification for importing a ton of code from Qt:

The root of the problem is that QListView uses *LVI::text(int column) const to
perform the incremental search, but Richard's code bypasses this data
altogether, so text() always returns an empty string.

You can fix this quite easily by adding the missing functions, but then the
list drawing is corrupted - text is rendered twice.

The reason for this is Richard's classes rely on Q|KLVI to draw their
backgrounds and treeview paraphernalia.  If text() returns something, they
render that as well.

The patch pulls all of the background drawing code into
Kopete::UI::ListView::Item, making it fully responsible for drawing itself.

The alternative, 2 line hack that sets the LVI's text to an empty
string before calling the parent's paintCell, and restores it before carrying
out the LVI's own paint routine, doesn't work because setText() makes QListView redraw and you end up busy-waiting.

svn path=/trunk/kdenetwork/kopete/; revision=330999