Some fix to the display icons. The display picture was not correctly refreshed because a bad check for KTempFile::file() wich is now set to 0L when the file is closed.
Also honour the size if icons of the toolbar.  To do this, i had to find a lot of workaround and hacks to get his size.
It also doesn't work if you close the chat window, recaive a message with the queue enabled, and open a chat with another contact bofore reading messages.  then the size is back to the default (22x22)

svn path=/trunk/kdenetwork/kopete/; revision=331299