Apply a different fix to bug #59080 which doesn't cause bug #89911.
Basically, this is a revert of Jason's commit of 28/05/04 which appears to
be bogus; dynamically adding/removing a toolbar is not fundamentally
different to adding/removing toolbar buttons or menu items -- we should not
do it. Plus the existing fix for bug #59080 disabled this behaviour anyway.

CCBUG: 59080
CCBUG: 89911
CCMAIL: matt@matt.rogers.name

svn path=/trunk/kdenetwork/kopete/; revision=356112