Rename Account::addContactToMetaContact to  createContact  (to be consistant with the Protocol::createAccount) Remove the displayName arg of this function.
Note: this function still return bool.  It would be nice if it could return a Contact*  anyway, in MSN, the contact is not created there, but later, whern the server reply. This is not realy needed, but is more beautifull

Can you please review the code i modified in your plugin ?

svn path=/trunk/kdenetwork/kopete/; revision=359000