I'm preparing to merge metacontact with the kopete_api_reword branch. The SyncMode flag from function to move the metacontact group has been removed.
This patch is a temporary (but working) replacement of it at Jabber level.
Till will be able to fix it in a proper way later. (thank)

(like i said, this is a simple replacement which do exactly the same thing than before, but per jabber account, and not for the whole libkopete.  idealy, this should probably be done by contact, but that's more difficult because some contact may be synced, and others not)

svn path=/trunk/kdenetwork/kopete/; revision=359946