Synced ChatView::raise and KopeteEmailWindow::raise. ChatView no longer calls KWin::raiseWindow since the contact list is not a pager. Kopete now seems to work correctly in Extreme focus protection mode.
Mr Backports: there's a small possibility that this won't work correctly on
KDE3.2. Correct behaviour is to do the sane thing if focus protection <=
High, and to do nothing (not even deiconify the window) if focus protection
is Extreme. Feel free not to backport this if you like.

BUG: 87416
CCMAIL: mattr@kde.org (Mr Backports)

svn path=/trunk/kdenetwork/kopete/; revision=363676