Add the API for plugins to access context menu events and tooltip events on HTML nodes in the chatview, as posted to the ML. No one had any other reccomendations on how it could be done, so I will take that as an "ok" sign.
Plugins can connect to the viewCreated signal of the ChatSessionManager in order to know when a view has been created.
They can then register themselves with that view ( if they want ) as handlers of tooltip or context menu events.

TODO: Not sure yet hwo to properly handle multiple plugins provifing a tooltip. Right now it would just use the last
one. Maybe they should be concatenated together?

TODO: Follow this shortly with a "Web Shortcuts" plugin that will let you right click on any word and run a shortcut on it.
Possibly also do a proof-of-concept tooltip plugin by using KDict to look up a word and put the result in the tooltip.

svn path=/trunk/kdenetwork/kopete/; revision=373079