BUG: 96052
Client like fire add <BR> instead of linebreak. So we need to escape them before passing them to pgp.

We only call Message::unsescape  if we detect a '<' in the message for performance.

Also, no need to do that for outgoing message, like the original patch proposed.

Thanks to Martin Albrecht for his patch.

svn path=/trunk/kdenetwork/kopete/; revision=377473