Make joining a channel non-blocking again, basically reverting richard's commit of Nov 4th.
Until we have truely async adding of contacts to the manager, and the ability to add contacts
in batches to speed things up and have less resorts, this hack will need to remain in place.
Better to have a bit slower overall joining of a large channel, thean to have it block the whole
GUI for tens of seconds.

Note this does not in any way fix this bug, but at least the GUI is no longer blocked while
this channel join is in place.

CCMAIL: 95077@bugs.kde.org

svn path=/trunk/kdenetwork/kopete/; revision=380309