Cleanup KABC.
- Add KABC export wizard that exports all or some metacontacts to new KABC
entries.
- Move as much KABC code as possible into a new singleton class,
KABCPersistence.  This contains a globally accessible addressbook reference,
which solves some consistency problems, and makes it possible to write the
wizard.
- Move the cryptic "Sync KABC" MCLVI* context menu entry to a "Import
contacts" button in the LVI props dialog.  This creates contacts on the MC
from IM addresses found in KABC, in case you never dared press it before.
- Rename the "Merge with addressbook" button in the same dialog to "Export
details".

One interesting change is that this patch no longer removes IM information
when changing metacontact associations.  The reason for this is that
Konversation also uses this information and it's not nice to remove other
programs' data.

CCMAIL: kopete-devel@kde.org

svn path=/trunk/kdenetwork/kopete/; revision=384571