When using the crypto plugin, store the message uncrypted in the history.
BUG: 77482

Having it encrypted is fully useless. You can't even read again message you sent because they are generally not encrypted with your own key (doing it may result in too big messages for some protocols)

If you want to keep the history secure:  1) don't log at all.  2) use an encrypted partition. so that it it secure.

svn path=/trunk/kdenetwork/kopete/; revision=387046