Fix the shutdown sequence on session exit by rewriting the way Kopete closes.
This patch borrows a bit of code from Kmail so that when the system tray icon is not shown, all kopete windows are responsible for keeping the app running, like Kmail.

There is still a problem that on session quit, the plugin manager does not fully complete its shutdown - the plugins are all shut down, but the KopeteApplication destructor is called before the plugins are deleteLater()'d.  Since the plugins are shutdown, Kopete no longer crashes on session exit.

See note in kopetewindow.cpp for more details.

svn path=/trunk/kdenetwork/kopete/; revision=397532