FEATURE: 106399
Support for global identity module.
What is Global Identity ?

It's a configuration module that set a global
identity(nickname, avatar) for all the registred accounts. It's support a global nickname you enter in the configuration or you can select a account where the global nickname will be sync.
Instead of changing your nickname for each account you have, only change the global nickname and it will be applied on each account that it supported !

Currently only implemented in MSN protocol. Jabber to coming soon !


svn path=/trunk/KDE/kdenetwork/kopete/; revision=418917