I have insisted Gof about this like ten times before. But it came to the third time I got all my 200 kabc hand-chosen pictures overwritten with stupid MSN avatars due to this broken logic.
You cannot assume that if kabcId is not null and picture is null that then
you don't have a picture and the best option is to sync the first one over it.
Why?
What happens if a kabc resource is temporary unavailable? it will satisfy that
condition, and you will lose the 200 pictures just like me.



svn path=/trunk/KDE/kdenetwork/kopete/; revision=425127