Refactor the displayName and photo properties, moving the current d->displayName semantics to a concept called custom displayname, and adding also kabc and subcontact as sources. In the past, if the subcontact value it was tracking was null, it didn't track and used kabc instead, and it wasn't using kabc at all, but a copy of it saved in a custom display name, a bit of mess.
This means, you can now select a subcontact to track its photo and name, and 
have the metacontact associated with a kabc entry, and also have a custom 
displayname, and select the source of the property without losing that 
information. No more sync, overwrites and complicated logic. If you use kabc 
source, then displayName() and photo() returns what is exactly contained in 
kabc. if you use custom source, those function returns the custom values, and 
if the source is contact, it will return the values right from the 
subcontact. 

The sync photo checkbox now has a clear semantic, it will sync the photo to 
kabc, no matter if it comes from contact (implemented) or from a custom one 
(not implemented yet), (if the source is kabc itself, the checkbox is 
disabled). And it means we could implement the same checkbox for the name.

Brief: it makes the stuff simpler and easy to understand. Forget sync back 
forward up down jump overwrite lose and import of properties.

The patch does a smart convertion from the old system. For example, it even 
checks the case you have a kabc id and you are not tracking a contact, and 
the custom displayname is the same as the kabc one, then it will set to track 
directly from kabc.

The metacontact properties dialog becomes more simpler to understand:
http://img99.echo.cx/img99/5898/mcdialog65yy.png


svn path=/trunk/KDE/kdenetwork/kopete/; revision=428673