Remove KMessageBox calls from MSN socket class. (socket class shout not have GUI calls, separate layers).
----------
BUG: 109326

Serialize the contactGUID, make sure that the contactGUID is saved (it's essential in MSNP11). The contactGUID property is still saved but it's better than notting ;)
Minor idioms fix for the add contact thing. 
Check if the contactGUID is not empty before sending the SBP command (cause connect/disconnect). This make sure because sometimes the contactGUID is not saved.


svn path=/trunk/KDE/kdenetwork/kopete/; revision=437457