Support of receiving webcam with MSN.
It's not finished yet (it is shown in a ugly window, and it doesn't close well the stream),  but it works.
I'm so happy i want to commit now.

libmimic depends of glib,  i need someone to help me to configure Makefiles in a way i can disable webcam if glib is not found



svn path=/branches/KDE/3.5/kdenetwork/kopete/; revision=451743