Fix Away status I brake with my previous commits.
Unfortunately we have to do the online status setting twice for the contact,
so the user will get two messages when someone goes away or comes back.
Examples:

(14:02:57) #  SomeUser is now Online.
(14:02:57) #  SomeUser is now Op.
(14:02:57) #  OtherUser is now Away.
(14:02:57) #  OtherUser is now Away.

I couldn't figure out any easy way to circumvent this.

When a user goes away, we first have to set the user's online status to plain
Away to get proper status in the Contact List (so channel specific modes such
as Operator are not wanted here). But then we have to set the contact's online
status again for all channels, to set the channel specific status to things
like Operator | Away. Both of these changes emit onlineStatusChanged(), which
will end up shown to the user.


svn path=/branches/work/kopete/dev-0.12/kopete/; revision=477227