add back the version of accounts that takes a Kopete::Protocol pointer.
Please keep API refactorings out of the Qt3 -> Qt4 porting process. Also,
please discuss API refactorings on the list first. libkopete is not a one
man show and I think that's it's only fair to the rest of us so we at least
know what's going on. 

CCMAIL: kopete-devel@kde.org

svn path=/trunk/KDE/kdenetwork/kopete/; revision=484831