Use a stack instead of a queue to open chatwindow.  and clear regulary this stack.
That mean the priority is given to the last contact you ask to chat to.

This fix Bug 118128: Wrong MSN contact when a window opens
BUG: 118128

Something which trigger this bug is also the fact that it need to download lots of picture when kopete start. disabling the automatic download of picture may help.
(this bug, with the combinaison of Bug 117089 may be really annoying.)




svn path=/branches/KDE/3.5/kdenetwork/kopete/; revision=487744