Try to fix bug 122269 CCMAIL: 122269
The problem is maybe because your icq transport is on your roster with a ressource, while kopete only expect bare Jid on the roster.
I've made Kopete more tollerent, and ignore the ressource.  I've also made it double check to don't add a transport twice (which is what may hapenned, Kopete has not detected the account existed because the contact has not the correct terminaison)

The ressounrce'd contact may have been added with an old version of kopete. I remember having fixed that because i had the problem with the icq gateway, preciselly.

Can you upgrade svn, and check if it still crash ?  if yes, please paste me the output since the kopete start



svn path=/branches/kopete/0.12/kopete/; revision=517722