Make the set status message menu work again. To enter a new message one now has to enter the string in in KInputDialog::getText() as it doesn't seem possible to put widgets in KMenus anymore.
Also fix Kopete::Away::getMessage(uint) : The requested message was removed from the
array of saved messages, now I've changed it back to the old behaviour that the returned
message is moved to the beginning of the array.

Can someone comment on the use of deleteLater() on KMenus? Some stuff was commented out
before; it appears that sometimes deleteLater() is used and sometimes it isn't.
I hope we do not leak memory here...

svn path=/trunk/KDE/kdenetwork/kopete/; revision=536124