Merged ChatSession from GroupWise plugin. Compiles, although there are tons of warning, and hundreds quick-n-dirty hacks and tens of FIXME and TODO. Runs OK, the framework works, the pseudo sendMessage works as well.
The conference mode is a little different for QQ protocol, I would introduce here just for the reference:

QQ does not support conference mode (more than 2 contacts talking) in the beginning, the conference mode is added via SIP( Qun/Group in QQ's termanology). The official QQ's behavior is like this:
A start a conversation with B, and A or B invites C to join, there is a new window popped out for A,B,C conversation, if D is also invited, D joins to the new Window.

From the packets sniffered, the conversation starts from the original sendIM/receiveIM, when A invites C, a SIP is created, and after that, all the conversation in the conference is in SIP mode.

I might stick to the universal behavior of Kopete, when A invites C, the private conversation between A & B is going to be destroyed.

First thing first, the next thing is cleaning the house, and implement the private conversation.


svn path=/trunk/KDE/kdenetwork/kopete/; revision=570992