Make sentence options (Capitalize and full stop.) work, and handle inbound messages as advertised.
NB Hariz Kouzinopoulos:  I did not use your patch, because while it is
technically correct, the job of saving the widgets' values is performed
automatically by KCAutoConfigModule, if it knows about the widgets.  A
previous author neglected to change the widgets passed to
KCAutoConfigModule when adding the sentence options, so that is why
these were not read/written.  Thank you for submitting a patch though,
it was the impetus to solve the bug properly.

BUG:116381

svn path=/branches/KDE/3.5/kdenetwork/kopete/; revision=602278