About 10-20% performance improvement for formatStyleKeywords.
This is important when lots of messages are added to the chat at the
same time, like when browsing history with alt-left/right, or when the
style changes, e.g. when someone is changing display picture.

Especially the latter is unfortunately a fairly common thing, and Kopete
freezes horribly if you have chats open with a person changing pic.

Approved by Michael.


svn path=/branches/KDE/3.5/kdenetwork/kopete/; revision=634872