Separate the model creation from the view class.
Use KopeteWindow to be the controlling object for both the view
and them model. Since we'll most likely being using proxy models
to implement the contact grouping schemes, we need to have an external
entity controlling what model the view uses. Besides, the view knowing about
it's model is just wrong anyways. :)

svn path=/branches/work/kopete/break-the-contactlist/; revision=713239