Fix crash on Kopete quit. Use K_GLOBAL_STATIC to manage the singleton lifetime.
This is the second time I've done the conversion. The first time didn't go so well,
so I ended up more closely following the example this time. This means that
Kopete::PluginManager::Private is now just Kopete::PluginManagerPrivate. It also means
that the constructor and destructor for PluginManager are empty and the functionality was
moved into the private class. All is working now, and the best part is there's no crash
on Kopete quit anymore. :)


svn path=/trunk/KDE/kdenetwork/kopete/; revision=717538