Rewrite the identity status widget to use a more recognisable list widget containing the accounts, and a pushbutton to change display picture.
I had some problems with this as Kopete::Identity is only superficially
integrated with the rest of libkopete, so it's not possible to know a
new account's real identity at the usual
Kopete::AccountManager::accountRegistered() signal.  As a safe solution
for 4.0 the identity status widget is hidden when this signal is
received.  After 4.0, Kopete::Identity will need to be properly
integrated.  

Still to do is to connect the widgets up to display photo changes in the 
config dialog and make sure display name changes are effective.


svn path=/trunk/KDE/kdenetwork/kopete/; revision=735548