Adds a much-requested feature: a way for protocols to give a hint to Kopete which form a chat should take.  This allows Kopete to give a chatwindow a size and members list visibility appropriate to the chat.
We've previously told users who requested this feature that it was not 
possible for consistency's sake, but I feel that this has handicapped our use 
of chatrooms/IRC in kopete.  This solves this dilemma
elegantly.

* Form is an attribute of Kopete::ChatSession
* 2 forms of chat: Small and Chatroom
* this affects the initial format of the chatwindow
*  Small windows and Chatroom windows layout and members list visibility
changes by the user are sticky and persisted independently

Additionally, the feature allows (optionally) a compact variant of the chat 
view style to be selected dependent on chat session form.  This is 
implemented by munging the relative path to the variant css.  
FunkyStyle.css -> _compact_FunkyStyle.css.  The rationale for this is that a 
large style with avatars is not appropriate to a high traffic channel.  This 
allows a compact variant to the main style, but that keeps has consistent 
theming, and works within the Adium chat spec

Notes on the patch:
* Everything defaults to the current Small behaviour! Even if applied as is, 
the patch is fail-safe!
* ...Except testbed, for testing and demonstration.  This requests Chatroom 
style windows.

Patch reviewed by Duncan and Boiko
CCMAIL: mueller@kde.org


svn path=/trunk/KDE/kdenetwork/kopete/; revision=736598