Add metacontact execution
Use the activated signal from QAbstractItemView to start chats with a
metacontact. Right now, we expose the MetaContact object directly. Haven't
decided yet whether I want to keep doing it this way or have the item abstract
it so the view doesn't have to worry about the MetaContact API changing.

svn path=/branches/work/kopete/dev-4.2/; revision=832619