Fix bug 172485 Kopete has full CPU load (freezes) when receiving a long line.
The message.setHtmlBody( message.parsedBody() ); which should cache the emoticons actually didn't cache it and it made it even worse because setHtml is much slower than parsedBody if we have many emoticons (html tags).

So this patch removes this line and adds caching into the parsedBody function. The great thing is that we don't have to call setHtml which was the bottleneck.

BUG: 172485


svn path=/trunk/KDE/kdenetwork/kopete/; revision=872704