BUG: 161536 getImage() was attempting to access elements in an empty vector. Have added a check at the start for the size of the vector to prevent this.
There is no UI feedback when the camera is already taken like this,
as it seems the camera still responds to some commands causing 
kopete to think it's available. Perhaps the UI should be modified
to show the camera is unavailable? The QImage could be changed at
this point in the code to display a camera unavailable image.

svn path=/trunk/KDE/kdenetwork/kopete/; revision=902727