Move this code into a K_GLOBAL_STATIC. There are two problems with this code:
The first is that it calls i18n() before KLocale and KComponentData
have initialised. That means we never get any translation.

The second is that it causes the QIconvCodec in Qt to initialise
before the locale system in libc is initialised (that's done inside
QCoreApplication's constructor). This means this thread is confined to
ASCII, thereby causing all X11 key translations to be broken.

This sounds like a solution to bug 182996, but my symptom was wildly
different: I had the problem in any and every text input. Please
retest the bug's symptom and close if it's fixed.

CCBUG:182996

svn path=/trunk/KDE/kdenetwork/kopete/; revision=922100