Changed deprecated uses of KTabWidget members:
* setHoverCloseButton(bool) -> setTabsClosable(bool)

* setTabReorderingEnabled(bool) -> SetMovable(bool)

Tabs now use the qt-provided methods of enabling dragging, and enabling the close button on them.

I don't know why the close button on the tabs themselves can't be enabled, it appears to be a config setting, but I couldn't find anything like it in the settings dialog, but this was like this before anyways.

svn path=/trunk/KDE/kdenetwork/kopete/; revision=933407