Rewrite the TelepahthyEditAccountWidget to reuse code from the Telepathy Accounts KCM.
At the moment, the main code from telepathy-accounts-kcm is copied into the accounts_kcm subdirectory of the telepathy protocol. This is a temporary measure until it's decided what the best way is to share the code between the two apps.

It is now necessary to build and install the KCM before the Kopete plugin.

There *are* many known bugs, which will get fixed in the near future, so please please don't expect this to work perfectly :).

Merge branch 'add-account-wizard-experimental-import'

svn path=/trunk/playground/network/kopete/protocols/telepathy/; revision=1007871