Fix kopete bonjour plugin handling of new contact added to existing metacontact.
When a bonjour contact becomes visible for whom a metacontact already exists, it would try and cast the first contact in the metacontact to BonjourContact even if it was for example a JabberContact, resulting in a segfault.
The fix uses mc->findContact to find the new contact without making assumptions about it's position inside the metacontact.

CCBUG: 177487
CCBUG: 192087

svn path=/trunk/KDE/kdenetwork/kopete/; revision=1034486