Video-settings (videodevicepool): fix workflow during device-open
Fixes
- loading of the saved device
- closing of the current device before opening a new device
- handling of multiple clients

With the current code, always the device with index 0 is used regardless of what has been saved.
This happens, because the video-settings are loaded AFTER the device is opened.
Because the video-settings must be applied with an open device, the only solution is to split the loadConfig()-function into two functions:
- loadSelectedDevice() => applied before the device is opened to set the right device-index
- loadDeviceConfig()   => applied after the device has been successfully opened to set the video-settings

With the current workflow during device-open (open(int device) calls open()), we do not know if the device is selected explicitly or the default device should be opened.
Because of that, merge open() and open(ind device), using negative parameters for selecting the standard/saved device (defualt=-1).
This way, we also fix handling of multiple clients for a device and closing of the previous device.

svn path=/branches/KDE/4.4/kdenetwork/kopete/; revision=1075262